feat: integrate immutable argument graph
- Added ImmutableArgumentGraph class to manage deep normalization and freezing of argument-style value graphs.
- Refactored ImmutableArguments and workflow step argument handling to utilize the new ImmutableArgumentGraph for normalization.
- Enhanced error handling in Catalog to raise InvalidDefinitionError for unknown workflow IDs.
- Updated specs to cover new functionality and ensure correctness of argument normalization and error handling.
