+++ core/karya/lib/karya/workflow/step.rb
@@ -5,6 +5,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+require_relative '../internal/immutable_argument_graph'
+
 module Karya
   module Workflow
     # Immutable one-step workflow composition unit.
@@ -22,98 +24,17 @@ module Karya
       # Normalizes workflow step arguments into the same immutable scalar graph
       # shape used by jobs without coupling workflow code to job internals.
       class Arguments
-        IMMUTABLE_SCALAR_CLASSES = [NilClass, Numeric, Symbol, TrueClass, FalseClass].freeze
-        DUPLICABLE_SCALAR_CLASSES = [String, Time].freeze
-
         def initialize(arguments)
           @arguments = arguments
         end
 
         def normalize
-          raise InvalidDefinitionError, 'arguments must be a Hash' unless arguments.is_a?(Hash)
-
-          normalize_hash(arguments, tracker: TraversalTracker.new)
+          Internal::ImmutableArgumentGraph.new(arguments, error_class: InvalidDefinitionError).normalize
         end
 
         private
 
         attr_reader :arguments
-
-        def normalize_hash(value, tracker:)
-          tracker.around(value) do
-            value.each_with_object({}) do |(key, item), normalized|
-              normalized_key = key.to_s.strip
-              raise InvalidDefinitionError, 'argument keys must be present' if normalized_key.empty?
-
-              normalized_key = normalized_key.freeze
-              if normalized.key?(normalized_key)
-                raise InvalidDefinitionError,
-                      "duplicate argument key after normalization: #{normalized_key.inspect}"
-              end
-
-              normalized[normalized_key] = normalize_value(item, tracker:)
-            end.freeze
-          end
-        end
-
-        def normalize_value(value, tracker:)
-          case value
-          when Hash
-            normalize_hash(value, tracker:)
-          when Array
-            normalize_array(value, tracker:)
-          else
-            normalize_scalar(value)
-          end
-        end
-
-        def normalize_array(value, tracker:)
-          tracker.around(value) do
-            value.map { |item| normalize_value(item, tracker:) }.freeze
-          end
-        end
-
-        def normalize_scalar(value)
-          arguments_class = self.class
-          return value if arguments_class.immutable_scalar?(value)
-          return value.dup.freeze if arguments_class.duplicable_scalar?(value)
-
-          raise InvalidDefinitionError,
-                'argument values must be composed of Hash, Array, String, Time, Symbol, Numeric, boolean, or nil'
-        end
-
-        class << self
-          def immutable_scalar?(value)
-            IMMUTABLE_SCALAR_CLASSES.any? { |klass| value.is_a?(klass) }
-          end
-
-          def duplicable_scalar?(value)
-            DUPLICABLE_SCALAR_CLASSES.any? { |klass| value.is_a?(klass) }
-          end
-        end
-
-        # Detects recursive argument graphs before they recurse forever.
-        class TraversalTracker
-          def initialize
-            @entered_object_ids = {}
-          end
-
-          def around(value)
-            object_id = value.object_id
-            raise InvalidDefinitionError, 'arguments must not contain recursive structures' if entered_object_ids.key?(object_id)
-
-            entered_object_ids[object_id] = true
-            yield
-          ensure
-            entered_object_ids.delete(object_id)
-          end
-
-          private
-
-          attr_reader :entered_object_ids
-        end
-
-        private_constant :TraversalTracker
       end
 
       # Normalizes one step's prerequisite list into frozen normalized ids.
